ITEP-81920 Add `Use Camera Pipeline` checkbox for Dynamic Camera Configuration (#618)
Purpose: UX improvement.

This PR enables the user to clearly indicate the intention whether to apply either camera settings or camera pipeline for deploying a video pipeline by:

adding Use Camera Pipeline checkbox in the Camera Calibration UI page with a help text explaining its purpose.
adding logic that applies the camera_pipeline field in the VA pipeline only if Use Camera Pipeline is checked.
updating user guide accordingly.
adding cross-field input validation in the Camera Calibration form and REST API serializer.
improving pipeline generation error handling in the form, so that only messages from known errors are shown in the UI.
Previously it was assumed that if Camera Pipeline field is not empty it means user's intention to use it instead of generating camera settings, which actually was wrong approach, because the user could have generated it only for the purpose of reviewing the pipeline string and subsequent edits to camera setting fields were not applied in such case.