ITEP-81806 Fix metric test hang-up (#625)
This PR:

fixes a race condition in time-chunked tracker that cause metrics tests to hang-up.
restores the metrics tests.
add some debug logs that may be useful in case of debugging tracker in future.
The race condition occurred at the end of the test in Tracking.join method, when time-chunking background thread put pending detections to the queue just after the tracker queue consumer thread has ended but before the tracker.queue.join() method was called. As a result, the queue.join() call blocked because there was no consumer to make the queue empty again.

The fix is implemented in overridden TimeChunkedIntelLabsTracking.join method which first waits for time-chunking background thread to drop all pending items and finish and only them calls Tracking.join.