ITEP-77684: UI allows empty camera and sensor creation (#624)
This PR allows camera and sensor scene field to be empty in UI. The behavior is now aligned with API.
It also auto fills the scene field based on origin: if coming from a scene, populate with that scene. If coming from cameras/sensors tab, the field is empty.

Additional fixes:
camera-status test fixed
sensor-area and sensor-scene moved out of broken-tests
mqtt-large-messages test removed